Updated Context Entities and Migrations, now working properly.

+++ PMStudio/Models/Entities/InstanciasPlantillasPasosDetalle.cs
@@ -11,11 +11,11 @@ namespace PMStudio.Models.Entities
             PasosUsuariosDetalle = new HashSet<PasosUsuariosDetalle>();
         }
 
-        public decimal InstanciaPlantilla { get; set; }
-        public decimal Paso { get; set; }
-        public decimal? Estado { get; set; }
-        public decimal IdPlantillaPasoDetalle { get; set; }
-        public decimal? UsuarioAccion { get; set; }
+        public int InstanciaPlantilla { get; set; }
+        public int Paso { get; set; }
+        public int? Estado { get; set; }
+        public int IdPlantillaPasoDetalle { get; set; }
+        public int? UsuarioAccion { get; set; }
         public string AspNetUser { get; set; }
 
         public virtual Acciones EstadoNavigation { get; set; }
