+++ PMStudio/Models/Entities/PasosInstanciasDatosDetalle.cs
@@ -5,10 +5,10 @@ namespace PMStudio.Models.Entities
 {
     public partial class PasosInstanciasDatosDetalle
     {
-        public decimal InstanciaPlantillaDato { get; set; }
-        public decimal Paso { get; set; }
+        public int InstanciaPlantillaDato { get; set; }
+        public int Paso { get; set; }
         public string SoloLectura { get; set; }
-        public decimal IdPasosInstanciasDatos { get; set; }
+        public int IdPasosInstanciasDatos { get; set; }
 
         public virtual InstanciasPlantillasDatosDetalle InstanciaPlantillaDatoNavigation { get; set; }
         public virtual PasosInstancias PasoNavigation { get; set; }
