+++ PMStudio/Models/Entities/PlantillasCamposDetalle.cs
@@ -5,8 +5,8 @@ namespace PMStudio.Models.Entities
 {
     public partial class PlantillasCamposDetalle
     {
-        public decimal IdPlantillaCampo { get; set; }
-        public decimal Plantilla { get; set; }
+        public int IdPlantillaCampo { get; set; }
+        public int Plantilla { get; set; }
         public string NombreCampo { get; set; }
 
         public virtual Plantillas PlantillaNavigation { get; set; }
