+++ PMStudio/Models/Entities/PlantillasPasosDetalle.cs
@@ -5,9 +5,9 @@ namespace PMStudio.Models.Entities
 {
     public partial class PlantillasPasosDetalle
     {
-        public decimal IdPlantillaPaso { get; set; }
-        public decimal Plantilla { get; set; }
-        public decimal Paso { get; set; }
+        public int IdPlantillaPaso { get; set; }
+        public int Plantilla { get; set; }
+        public int Paso { get; set; }
 
         public virtual Pasos PasoNavigation { get; set; }
         public virtual Plantillas PlantillaNavigation { get; set; }
